Working on edit component
